Added S3 bucket support
Attempting to mount an S3 account will connect the mount to ALL buckets on associated AWS account, which is undesirable in most cases. In fact, I can't get rclone to actually see files in individual buckets when it mounts the entire account (this particular problem is an upstream issue and not yours!).

This PR adds the ability to specify a specific bucket, which will mount only the desired bucket while also bypassing the upstream issue with mounting S3 accounts. I'm currently rolling the above changes on my own phone without any apparent problems.